Entityactions (request) (#452)
* entityactions

* entityactions II

* moved method
