+++ src/api/java/de/teamlapen/vampirism/api/entity/actions/IEntityActionUser.java
@@ -1,7 +1,6 @@
 package de.teamlapen.vampirism.api.entity.actions;
 
-import com.google.common.collect.Lists;
-
+import de.teamlapen.vampirism.api.VampirismAPI;
 import de.teamlapen.vampirism.api.difficulty.IAdjustableLevel;
 import de.teamlapen.vampirism.api.entity.EntityClassType;
 import de.teamlapen.vampirism.api.entity.factions.IFactionEntity;
@@ -22,6 +21,6 @@ public interface IEntityActionUser extends IAdjustableLevel, IFactionEntity {
      * gets all available actions for this entity.
      */
     default List<IEntityAction> getAvailableActions() {
-        return Lists.newArrayList();
+        return VampirismAPI.entityActionManager().getAllEntityActionsByTierAndClassType(this.getFaction(), getEntityTier(), getEntityClass());
     }
 }
